Reversed value of B-select in the ControlWord to reflect the orientation of the multiplexer in the ALU
The B_Select employed before assumed that when it was high, we would use the constant fed in by the user. However the ALU did not operate this way, and would use a constant when B_Select was low. This commit reflects this by reversing the value of B_select for all cases.